Split annotations-objects.ts to many files
