+++ cvat-core/src/annotations-objects/shape-models.ts
@@ -1,0 +1,13 @@
+// Copyright (C) 2019-2022 Intel Corporation
+// Copyright (C) CVAT.ai Corporation
+//
+// SPDX-License-Identifier: MIT
+
+export { RectangleShape } from './rectangle-shape';
+export { EllipseShape } from './ellipse-shape';
+export { PolygonShape } from './polygon-shape';
+export { PolylineShape } from './polyline-shape';
+export { PointsShape } from './points-shape';
+export { CuboidShape } from './cuboid-shape';
+export { SkeletonShape } from './skeleton-shape';
+export { MaskShape } from './mask-shape';
