+++ cvat-core/src/annotations-objects/track-models.ts
@@ -1,0 +1,12 @@
+// Copyright (C) 2019-2022 Intel Corporation
+// Copyright (C) CVAT.ai Corporation
+//
+// SPDX-License-Identifier: MIT
+
+export { RectangleTrack } from './rectangle-track';
+export { EllipseTrack } from './ellipse-track';
+export { PolygonTrack } from './polygon-track';
+export { PolylineTrack } from './polyline-track';
+export { PointsTrack } from './points-track';
+export { CuboidTrack } from './cuboid-track';
+export { SkeletonTrack } from './skeleton-track';
